requirements.txt
fastapi[all]~=0.115.6
requests
pydantic~=2.10.4
uvicorn~=0.34.0
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1